Move Reader base class to api module
This allows completing the type annotations for the load function.

Tests were also fixed guided by the new annotations to cover all cases
of the Chunk reader constructor.
